Moved OnButtonListener to manager package
