+++ base/src/main/java/com/smartdevicelink/managers/screen/BaseScreenManager.java
@@ -37,6 +37,7 @@ import android.util.Log;
 
 import com.smartdevicelink.managers.BaseSubManager;
 import com.smartdevicelink.managers.CompletionListener;
+import com.smartdevicelink.managers.OnButtonListener;
 import com.smartdevicelink.managers.file.FileManager;
 import com.smartdevicelink.managers.file.filetypes.SdlArtwork;
 import com.smartdevicelink.managers.screen.choiceset.ChoiceCell;
@@ -55,7 +56,6 @@ import com.smartdevicelink.proxy.rpc.enums.ButtonName;
 import com.smartdevicelink.proxy.rpc.enums.InteractionMode;
 import com.smartdevicelink.proxy.rpc.enums.MetadataType;
 import com.smartdevicelink.proxy.rpc.enums.TextAlignment;
-import com.smartdevicelink.managers.screen.OnButtonListener;
 import com.smartdevicelink.util.DebugTool;
 
 import java.lang.ref.WeakReference;
@@ -596,10 +596,10 @@ abstract class BaseScreenManager extends BaseSubManager {
 		});
 	}
 
-	public void addButtonListener(ButtonName buttonName, OnButtonListener listener){
+	public void addButtonListener(@NonNull ButtonName buttonName, @NonNull OnButtonListener listener){
 		subscribeButtonManager.addButtonListener(buttonName,listener);
 	}
-	public void removeButtonListener(ButtonName buttonName, OnButtonListener listener){
+	public void removeButtonListener(@NonNull ButtonName buttonName, @NonNull OnButtonListener listener){
 		subscribeButtonManager.removeButtonListener(buttonName, listener);
 	}
 }
